requirements.txt

Flask-CORS==4.0.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
gunicorn==21.2.0
Werkzeug==2.3.7
Flask==3.0.0
python-dotenv==1.0.0
dateutils==0.6.12
Flask-Limiter>=3.0.0
google-generativeai==0.3.0
